Make DNS smoketest work when admin has a public IP.
It is possible that the IP for a system reported by knife is not the
one we want to use -- knife always reports the IP on a network with a
gateway if it can, and DNS is hardwired to always talk over the admin
network.

For now, assume that the DNS tests pass on the clients if we can
nslookup the DNS server.