Convert the attribute for records to a hash instead of referencing it
Since we are putting the final zone mash in the attribute, if we just
reference the records, then we will save a reference to the records
instead of something different.

This will be an issue soon, since we will compare the zones from the
attribute to the zone we're rebuilding to see if there's a change.