Make dnsmasq not rely on external /etc/resolv-forwarders.conf file
The settings can very well be in dnsmasq.conf (for which we have a
template), and when apparmor is used, dnsmasq can't read that file.

That also means that we need to use the no-resolv option to not have
dnsmasq use /etc/resolv.conf (since we don't override that file
anymore).