bind9: Remove unused reverse IP zones
When a node is deleted or an IP address is deallocated, nothing in the
bind config references the reverse IP zones for the IP addresses that
are removed. But the files still stay in the directory, which can be
confusing.

So let's drop them!

Backport of https://github.com/crowbar/crowbar-core/pull/98